requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
alembic==1.13.3
annotated-types==0.7.0
anthropic==0.37.1
anyio==4.6.2.post1
appdirs==1.4.4
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
auth0-python==4.7.2
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.7
cohere==5.11.1
coloredlogs==15.0.1
crewai==0.76.2
crewai-tools==0.13.2
cryptography==43.0.3
cssselect==1.2.0
dataclasses-json==0.6.7
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
deprecation==2.1.0
distro==1.9.0
docker==7.1.0
docstring_parser==0.16
docx2txt==0.8
durationpy==0.9
embedchain==0.1.123
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.3
fastavro==1.9.7
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.16.1
flatbuffers==24.3.25
frozenlist==1.5.0
fsspec==2024.10.0
google-api-core==2.21.0
google-auth==2.35.0
google-cloud-aiplatform==1.70.0
google-cloud-bigquery==3.26.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.5
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
gptcache==0.1.44
greenlet==3.1.1
grpc-google-iam-v1==0.13.1
grpcio==1.67.0
grpcio-status==1.62.3
grpcio-tools==1.62.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.1
humanfriendly==10.0
hyperframe==6.0.1
idna==3.10
importlib_metadata==8.4.0
importlib_resources==6.4.5
iniconfig==2.0.0
instructor==1.6.3
ipython==8.28.0
jedi==0.19.1
jieba3k==0.35.1
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
json_repair==0.30.0
jsonpatch==1.33
jsonpickle==3.3.0
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kubernetes==31.0.0
lancedb==0.14.0
langchain==0.3.4
langchain-anthropic==0.2.3
langchain-cohere==0.3.1
langchain-community==0.3.3
langchain-core==0.3.12
langchain-experimental==0.3.2
langchain-openai==0.2.3
langchain-text-splitters==0.3.0
langgraph==0.2.39
langgraph-checkpoint==2.0.2
langgraph-sdk==0.1.33
langsmith==0.1.137
litellm==1.50.4
lxml==5.3.0
Mako==1.3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mem0ai==0.1.23
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.4.2
newsapi-python==0.2.7
newspaper3k==0.2.8
nltk==3.9.1
nodeenv==1.9.1
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.19.2
openai==1.52.2
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
orjson==3.10.10
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.1
pandas==2.2.3
parameterized==0.9.0
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
pluggy==1.5.0
portalocker==2.10.1
posthog==3.7.0
prompt_toolkit==3.0.48
propcache==0.2.0
proto-plus==1.25.0
protobuf==4.25.5
ptyprocess==0.7.0
pulsar-client==3.5.0
pure_eval==0.2.3
py==1.11.0
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
pylance==0.18.2
pypdf==5.0.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyright==1.1.386
pysbd==0.3.4
PySocks==1.7.1
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytube==15.0.0
pytz==2024.2
pyvis==0.3.2
PyYAML==6.0.2
qdrant-client==1.12.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
retry==0.9.2
rich==13.9.3
rpds-py==0.20.0
rsa==4.9
schema==0.7.7
scikit-learn==1.5.2
scipy==1.14.1
selenium==4.25.0
sgmllib3k==1.0.0
shapely==2.0.6
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.36
stack-data==0.6.3
starlette==0.41.0
sympy==1.13.3
tabulate==0.9.0
tavily-python==0.5.0
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tinysegmenter==0.3
tldextract==5.1.2
tokenizers==0.20.1
tomli==2.0.2
tomli_w==1.1.0
tqdm==4.66.5
traitlets==5.14.3
trio==0.27.0
trio-websocket==0.11.1
typer==0.12.5
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uv==0.4.26
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==13.1
wrapt==1.16.0
wsproto==1.2.0
yarl==1.16.0
zipp==3.20.2
